refactor: #385: split position store into independent functions
